feat: Use Devs Deps analyzer as base class for supported analyzers
